Fix compile warnings
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@1209 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/exif.c
@@ -819,7 +819,7 @@ process_EXIF(unsigned char * const ExifData,
         if (ShowTags){
             fprintf(stderr, 
                     "Non-settings part of Exif header: %lu bytes\n",
-                    ExifData+length-LastExifRefd);
+                    (unsigned long)(ExifData+length-LastExifRefd));
         }
     }
 }
